requirements.txt
Flask==0.10.1
Flask-SocketIO==0.5.0
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.5
gunicorn==18.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
Werkzeug==0.10.1
